backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
python-multipart==0.0.9
google-generativeai==0.4.1
openai==1.17.0
sqlalchemy==2.0.29
pdfplumber==0.11.0
python-docx==1.1.0
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
